Fix crafting goalFilters: dynamicaly detect settings, make inclusive
